backend/requirements.txt
algoliasearch==2.6.1
algoliasearch-django==2.0.0
aniso8601==9.0.1
APScheduler==3.6.3
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
cachetools==4.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.0
cloudinary==1.29.0
cookiecutter==1.7.3
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-browser-reload==1.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-heroku==0.3.1
django-tailwind==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jinja2-time==0.2.0
MarkupSafe==2.1.1
Pillow==9.1.0
platformdirs==2.5.1
poyo==0.5.0
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-slugify==6.1.1
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
urllib3==1.26.9
websockets==10.2
Werkzeug==2.1.0
whitenoise==6.0.0
